# src/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "Documentacion de la API de webinar '_Webinar Organizador de Tareas con autenticación_' realizado en [Express](https://expressjs.com/) con el uso del ORM [Prisma](https://www.prisma.io/) usando una base de datos en [PostgreSQL](https://postgresql.com/)",
    "version": "1.0.0",
    "title": "Webinar Organizador de Tareas con autenticación",
    "termsOfService": "http://swagger.io/terms/",
    "contact": {
      "email": "ederiveroman@gmail.com"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
  "host": "webinar-task-codigo.herokuapp.com",
  "basePath": "/api/v1",
...
  "schemes": [
    "https"
  ],
  "paths": {
...
    "/task/{id}": {
      "get": {
        "tags": [
          "tasks"
        ],
        "summary": "Devuelve una task",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Id de la task",
            "required": true,
            "type": "string"
          }
        ],
        "operationId": "task",
        "responses": {
          "200": {
            "description": "La task encontrada",
            "schema": {
              "$ref": "#/definitions/TaskResponse"
            }
          }
        },
        "security": [
          {
            "api_key": []
          }
        ]
      },
      "patch": {
        "tags": [
          "tasks"
        ],
        "summary": "Actualiza parte de la task",
        "operationId": "patchTask",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Credenciales del usuario",
            "required": true,
            "type": "string"
          },
          {
            "in": "body",
            "name": "body",
            "description": "Credenciales del usuario",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Task"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Task actualizada",
            "schema": {
              "$ref": "#/definitions/TaskResponse"
            }
          }
        },
        "security": [
          {
            "api_key": []
          }
        ]
      },
      "delete": {
        "tags": [
          "tasks"
        ],
        "summary": "Elimina una task",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Id de la task",
            "required": true,
            "type": "string"
          }
        ],
        "operationId": "deleteTask",
        "responses": {
          "204": {
            "description": "Task eliminada"
          }
        },
        "security": [
          {
            "api_key": []
          }
        ]
      }
    },
...
    "/attachment/{id}": {
      "delete": {
        "tags": [
          "attachments"
        ],
        "summary": "Eliminar el attachment del bucket y de la base de datos",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Id de la task",
            "required": true,
            "type": "string"
          }
        ],
        "operationId": "deleteAttachment",
        "responses": {
          "204": {
            "description": "Fue exitoso"
          }
        },
        "security": [
          {
            "api_key": []
          }
        ]
      }
    }
  },
  "securityDefinitions": {
    "api_key": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header",
      "description": "Bearer <YOUR_TOKEN>"
    }
  },
  "definitions": {
...
    "Task": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string"
        },
        "status": {
          "type": "string",
          "enum": [
            "TODO",
            "DOING",
            "DONE"
          ]
        }
      }
    },
    "TaskResponse": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string"
        },
        "name": {
          "type": "string"
        },
        "status": {
          "type": "string",
          "enum": [
            "TODO",
            "DOING",
            "DONE"
          ]
        },
        "attachment": {
          "type": "object",
          "properties": {
            "id": {
              "type": "string"
            },
            "path": {
              "type": "string"
            },
            "key": {
              "type": "string"
            },
            "ext": {
              "type": "string"
            },
            "contentType": {
              "type": "string"
            },
            "taskId": {
              "type": "string"
            },
            "createdAt": {
              "type": "string"
            },
            "signedUrl": {
              "type": "string"
            }
          }
        }
      }
    },
...
  },
...
}